requirements.txt
Django==1.4.2
django-annoying==0.7.6
httplib2==0.7.6
urllib3==1.5
wsgiref==0.1.2
